+++ packages/SwingSet/misc-tools/replace-bundle.js
@@ -3,7 +3,7 @@
 import '@endo/init';
 import process from 'process';
 import { openSwingStore } from '@agoric/swing-store';
-import bundleSource from '@agoric/bundle-source';
+import bundleSource from '@endo/bundle-source';
 
 const log = console.log;