+++ packages/SwingSet/misc-tools/replay-transcript.js
@@ -6,7 +6,7 @@ import zlib from 'zlib';
 import readline from 'readline';
 import process from 'process';
 import { spawn } from 'child_process';
-import bundleSource from '@agoric/bundle-source';
+import bundleSource from '@endo/bundle-source';
 import { waitUntilQuiescent } from '../src/waitUntilQuiescent.js';
 import { makeStartXSnap } from '../src/controller.js';
 import { makeXsSubprocessFactory } from '../src/kernel/vatManager/manager-subprocess-xsnap.js';