+++ packages/run-protocol/test/vaultFactory/test-bootstrapPayment.js
@@ -7,7 +7,7 @@ import '../../src/vaultFactory/types.js';
 
 import path from 'path';
 import { E } from '@agoric/eventual-send';
-import bundleSource from '@agoric/bundle-source';
+import bundleSource from '@endo/bundle-source';
 import { makeFakeVatAdmin } from '@agoric/zoe/tools/fakeVatAdmin.js';
 import { makeZoeKit } from '@agoric/zoe';
 import buildManualTimer from '@agoric/zoe/tools/manualTimer.js';