+++ packages/run-protocol/test/vaultFactory/test-vault-interest.js
@@ -6,7 +6,7 @@ import { E } from '@agoric/eventual-send';
 import { makeFakeVatAdmin } from '@agoric/zoe/tools/fakeVatAdmin.js';
 import { makeLoopback } from '@endo/captp';
 import { makeZoeKit } from '@agoric/zoe';
-import bundleSource from '@agoric/bundle-source';
+import bundleSource from '@endo/bundle-source';
 import { resolve as importMetaResolve } from 'import-meta-resolve';
 
 import { makeRatio } from '@agoric/zoe/src/contractSupport/ratio.js';