+++ packages/swingset-runner/demo/swapBenchmark/prepareContracts.js
@@ -1,6 +1,6 @@
 import '@endo/init/pre-bundle-source.js';
 import '@endo/init';
-import bundleSource from '@agoric/bundle-source';
+import bundleSource from '@endo/bundle-source';
 import { resolve as importMetaResolve } from 'import-meta-resolve';
 
 import fs from 'fs';