+++ packages/zoe/test/swingsetTests/zoe/test-zoe.js
@@ -9,7 +9,7 @@ import '@endo/init';
 import test from 'ava';
 import path from 'path';
 import { buildVatController, buildKernelBundles } from '@agoric/swingset-vat';
-import bundleSource from '@agoric/bundle-source';
+import bundleSource from '@endo/bundle-source';
 
 const filename = new URL(import.meta.url).pathname;
 const dirname = path.dirname(filename);