+++ packages/zoe/test/unitTests/contracts/loan/helpers.js
@@ -7,7 +7,7 @@ import path from 'path';
 import '../../../../exported.js';
 
 import { E } from '@agoric/eventual-send';
-import bundleSource from '@agoric/bundle-source';
+import bundleSource from '@endo/bundle-source';
 import { AmountMath } from '@agoric/ertp';
 
 import { setup } from '../../setupBasicMints.js';