+++ packages/zoe/test/unitTests/installFromSource.js
@@ -1,6 +1,6 @@
 // @ts-check
 
-import bundleSource from '@agoric/bundle-source';
+import bundleSource from '@endo/bundle-source';
 import { E } from '@agoric/eventual-send';
 
 /**