+++ packages/SwingSet/NEWS.md
@@ -4,7 +4,7 @@ User-visible changes in SwingSet:
 
 * change Command Device (`buildCommand()`) to take `broadcastCallback` as an
   argument, rather than set in a subsequent call (#185)
-* outsource vat/kernel source bundling with @agoric/bundle-source
+* outsource vat/kernel source bundling with @endo/bundle-source
 
 
 ## Release 0.2.0 (27-Oct-2019)